Добавлено описание класса и его методов
