+++ bs4scraper_sinc.py.patch
@@ -1,0 +1,72 @@
+# Patch generated by Pyment v0.3.3
+
+--- a/bs4scraper_sinc.py
++++ b/bs4scraper_sinc.py
+@@ -6,23 +6,22 @@
+ 
+ 
+ class ParserBS4:
+-    """
+-    Класс для парсинга сайта с помощью BeautifulSoup4.
+-
++    """Класс для парсинга сайта с помощью BeautifulSoup4.
++    
+     params: url: str: url сайта.
+     params: session: requests.Session: объект для работы с сайтом.
+     params: base_url: str: domain сайта.
+     params: category_href: bool: Флаг необходимости получения ссылки на категорию.
+-    params: item_href: bool: Флаг необходимости обработки карточки товара.
+ 
+-    methods:
+-    get_html(self) -> str: метод для получуения html сайта.
+-    get_soup(self, url:str = None) -> BeautifulSoup: метод для получения обьекты bs4.soup html сайта.
+-    get_category(self): -> str: метод для получения категории сайта.
+-    get_page(self) -> int: метод для получения номера страницы сайта.
+-    get_items(self) -> ResultSet: метод для получения элементов сайта.
+-    inspect_item(self) -> Iterable: метод для получения информации об обьекте.
+-    save_item(self) -> None: метод для сохранения информации об обьекте.
++    :param methods: 
++    :param get_html: self
++    :param get_soup: self
++    :param get_category: self
++    :param get_page: self
++    :param get_items: self
++    :param inspect_item: self
++    :param save_item: self
++
+     """
+ 
+     def __init__(self, url: str, category_href: bool = True, item_href: bool = True):
+@@ -35,9 +34,9 @@
+     def get_html(self, url: str) -> str:
+         """Метод для получения html сайта
+ 
+-        params: url: str: Url сайта.
++        :param url: str: 
+ 
+-        return: str: html сайта."""
++        """
+         response = self.session.get(url=url,
+                                     headers={'User-Agent': UserAgent().random},
+                                     timeout=5)
+@@ -48,8 +47,9 @@
+         """Метод для получения обьекты bs4.soup html сайта
+ 
+         :param url: str:  (Default value = None) Url сайта.
++        :param url: str:  (Default value = None)
++        :returns: BeautifulSoup: обьект bs4.soup
+ 
+-        :return: BeautifulSoup: обьект bs4.soup
+         """
+         if url is None:
+             url = self.url
+@@ -81,8 +81,7 @@
+                 yield item
+ 
+     def inspect_item(self) -> Iterable:
+-        """Метод для получения информации об обьекте
+-        """
++        """Метод для получения информации об обьекте"""
+         if self.item_href:
+             for item_page in self.get_item():
+                 soup: BeautifulSoup = self.get_soup(self.base_url + item_page)
